fix random strategy

+++ strategies/custom0.js
@@ -17,7 +17,7 @@ var strat = {};
 // Prepare everything our method needs
 strat.init = function() {
   this.input = 'candle';
-  this.currentTrend = 'long';
+  this.currentTrend = 'short';
   this.requiredHistory = 0;
   this.candles=[];
 }
